macos changes

+++ MoonRuntime/Source/ECS/Entity.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 // A simple type alias
-using Entity = size_t;
+using Entity = unsigned long long;
 
 // Used to define the size of arrays later on
 const Entity MAX_ENTITIES = 10000;
